==> 5_analyses_selection_spread_asymmetry.Rmd <==
---
title: "Selection favors high spread and asymmetry of flower opening dates within plant individuals"
author: "Alicia Valdés"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: yes
    toc_depth: '4'
    df_print: paged
  pdf_document:
    toc: yes
    toc_depth: 4
  word_document:
    toc: yes
    toc_depth: '4'
subtitle: "Analyses"
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
options(tibble.width = Inf)
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r load packages, include=FALSE}
library(tidyverse)
library(ggeffects)
library(glmmTMB)
library(DHARMa)
library(gridExtra)
library(ggthemes)
library(ggfortify)
library(ggcorrplot)
library(PearsonDS)
library(car)
library(broom.mixed)
library(performance)
library(sjPlot)
library(lme4)
library(purrr)
library(ggridges)
library(moments)
library(grid)
library(ggpubr)
library(plotrix)
library(brms)
library(bayesplot)
library(lmerTest)
library(emmeans)
```

```{r Define ggplot themes and palettes, include=FALSE}
my_theme <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(legend.position="none")+theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
my_theme_legend <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
```

```{r load previously created output}
load(file="output/models/brms_87.rda")
load(file="output/models/brms_88.rda")
load(file="output/models/brms_89.rda")
load(file="output/models/brms_allyears.rda")
load(file="output/b_sel_nfl_linear.RData")
load(file="output/BCIs_sel_linear_87.RData")
load(file="output/BCIs_sel_linear_88.RData")
load(file="output/BCIs_sel_linear_89.RData")
load(file="output/b_abs_nfl_linear.RData")
load(file="output/BCIs_abs_nfl_linear_87.RData")
load(file="output/BCIs_abs_nfl_linear_88.RData")
load(file="output/BCIs_abs_nfl_linear_89.RData")
load(file="output/b_sel_linear.RData")
load(file="output/BCIs_sel_nfl_linear_87.RData")
load(file="output/BCIs_sel_nfl_linear_88.RData")
load(file="output/BCIs_sel_nfl_linear_89.RData")
load(file="output/permANOVA_sel_nfl_linear.RData")
load(file="output/permANOVA_abs_nfl_linear.RData")
load(file="output/sel_linear.RData")
```

# Read clean data from .csv files

```{r}
data_ids <- read_csv("data/clean/data_ids.csv")
data_ids$imp_seed_preyed <- as.factor(data_ids$imp_seed_preyed)
data_ids$year <- as.factor(data_ids$year)
data_ids$subplot<-as.factor(data_ids$subplot)
data_id_flowers <- read_csv("data/clean/data_id_flowers.csv")
data_weather<-read_csv("C:/Users/alici/Dropbox/SU/Projects/lathyrus/lathyrus_ms1/data/clean/weather_GDD.csv")
```

# Data manipulation

```{r}
data_ids<-data_ids%>%
  dplyr::select(-n_fl)%>%
  rename(n_fl=n_fl_a)
```

```{r}
data_ids<-data_ids%>%
  dplyr::select(-n_fr) # Use n_mat_intact_fr 
```

# Recalculate fruit set

```{r}
data_ids<-data_ids%>%
  mutate(n_mat_intact_fr=ifelse(is.na(n_mat_intact_fr),0,n_mat_intact_fr))
```

```{r}
data_ids<-data_ids%>%
  mutate(fr_set=n_mat_intact_fr/n_fl)
```

# Ranges, means and SD of the four moments

```{r}
data_ids%>%
  summarise(min_mean=min(avFD_v),max_mean=max(avFD_v),
            mean_mean=mean(avFD_v),sd_mean=sd(avFD_v),
            min_var=min(var,na.rm=T),max_var=max(var,na.rm=T),
            mean_var=mean(var,na.rm=T),sd_var=sd(var,na.rm=T),
            min_skew=min(skew,na.rm=T),max_skew=max(skew,na.rm=T),
            mean_skew=mean(skew,na.rm=T),sd_skew=sd(skew,na.rm=T),
            min_kurt=min(kurt,na.rm=T),max_kurt=max(kurt,na.rm=T),
            mean_kurt=mean(kurt,na.rm=T),sd_kurt=sd(kurt,na.rm=T))
```

# Appendix 3: Flowering phenology curves of individuals

```{r}
ids_years_1_2fl<-data_id_flowers%>%group_by(year,id)%>%mutate(count=n())%>%
  filter(count==1|count==2)%>%dplyr::select(id,year)%>%distinct(id,year) 
# All ids and years with 1 or 2 fl

data_id_flowers_3fl<-data_id_flowers %>%
  anti_join(ids_years_1_2fl)%>% # Keep only ids and years with at least 3 fl
  mutate(subplot=factor(subplot,ordered=T),number=factor(number,ordered=T))%>%
  mutate(year=factor(year),
         id=factor(id,
                   levels=unique(id[order(subplot,number)]), ordered=TRUE))
# id as ordered factor based on subplot and number

# split the data into chunks of 21 subplot, number
# based on their unique id values
id_chunks <-data_id_flowers_3fl %>% 
  distinct(subplot,number) %>%
  arrange(subplot,number)%>%
  mutate(chunk = ceiling(row_number()/21)) %>%
  inner_join(data_id_flowers_3fl, by =c("subplot","number")) %>%
  split(.$chunk)

# create a list of plots, one for each chunk of ids
plot_list<-map(id_chunks,~ggplot(.x,aes(x=opening_date_v,y=id,fill=year))+
                   geom_density_ridges()+
                   theme_ridges()+theme(text=element_text(family="serif"))+
                   theme(legend.position="top")+
                   labs(x="Opening date (number of days from the vernal equinox)",
                        y="Individual")+
                   scale_fill_manual(values=c("#E69F00","#56B4E9","#009E73")))

# save the plots as TIFF files
walk2(plot_list, seq_along(plot_list), 
      ~ ggsave(paste0("output/figures/Appendix3_", .y, ".tiff"), 
               plot = .x, width = 6.5, height = 9, units = "in", dpi = 300))
```

# Figure 1

```{r}
curves_mean<-data.frame(
  low=rpearson(10000,
               moments=c(mean=60.1,variance=0.5,skewness=0,kurtosis=3)),
  medium=rpearson(10000,
                  moments=c(mean=66.2,variance=0.5,skewness=0,kurtosis=3)),
  high=rpearson(10000,
                moments=c(mean=72.4,variance=0.5,skewness=0,kurtosis=3)))%>%
  pivot_longer(cols=low:high,names_to="curve_id",values_to="value")%>%
  mutate(moment="Mean")
curves_var<-data.frame(
  low=rpearson(10000,
               moments=c(mean=66.2,variance=0.05,skewness=0,kurtosis=3)),
  medium=rpearson(10000,
                  moments=c(mean=66.2,variance=0.5,skewness=0,kurtosis=3)),
  high=rpearson(10000,
                moments=c(mean=66.2,variance=1,skewness=0,kurtosis=3)))%>%
  pivot_longer(cols=low:high,names_to="curve_id",values_to="value")%>%
  mutate(moment="Variance")
curves_skew<-data.frame(
  low=rpearson(10000,
               moments=c(mean=66.2,variance=0.5,skewness=-1.1,kurtosis=3)),
  medium=rpearson(10000,
                  moments=c(mean=66.2,variance=0.5,skewness=0,kurtosis=3)),
  high=rpearson(10000,
                moments=c(mean=66.2,variance=0.5,skewness=1.1,kurtosis=3)))%>%
  pivot_longer(cols=low:high,names_to="curve_id",values_to="value")%>%
  mutate(moment="Skewness")
curves_kurt<-data.frame(
  low=rpearson(10000,
               moments=c(mean=66.2,variance=0.5,skewness=0,kurtosis=2)),
  medium=rpearson(10000,
                  moments=c(mean=66.2,variance=0.5,skewness=0,kurtosis=3)),
  high=rpearson(10000,
                moments=c(mean=66.2,variance=0.5,skewness=0,kurtosis=4)))%>%
  pivot_longer(cols=low:high,names_to="curve_id",values_to="value")%>%
  mutate(moment="Kurtosis")
curves<-rbind(curves_mean,curves_var,curves_skew,curves_kurt)%>%
  mutate(moment=factor(moment,
                       levels=c("Mean","Variance","Skewness","Kurtosis")),
         curve_id=factor(curve_id,
                         levels=c("low","medium","high")))
```

```{r}
curves%>%group_by(moment,curve_id)%>%
  summarise(mean=mean(value),var=var(value),
            skew=skewness(value),kurt=kurtosis(value))
```

```{r}
fig1A<-grid.arrange(
  ggplot(curves%>%filter(moment=="Mean")%>%
           mutate(curve_id=factor(curve_id,levels=c("low","medium","high"))),
         aes(x=value,fill=curve_id,color=curve_id))+
    geom_density(alpha=0.5)+my_theme_legend()+theme(legend.position="top")+
    labs(fill="Mean",color="Mean")+
    xlab("Opening date")+ylab("Density")+ggtitle("A)")+
    scale_fill_manual(values=c("#2c7bb6","#ffeda0","#d7191c"),
                      labels=c("60.1", "66.2", "72.4"))+
    scale_color_manual(values=c("#265a93","#ffdd87","#b71416"),
                      labels=c("60.1", "66.2", "72.4"))+
    theme(plot.title = element_text(hjust =-0.09,vjust=-6))+ylab(NULL)+
    scale_x_continuous(breaks=c(60,62,64,66,68,70,72)),
  ggplot(curves%>%filter(moment=="Variance")%>%
           mutate(curve_id=factor(curve_id,levels=c("low","medium","high"))),
         aes(x=value,fill=curve_id,color=curve_id))+
    geom_density(alpha=0.5)+my_theme_legend()+theme(legend.position="top")+
    labs(fill="Variance",color="Variance")+
    xlab("Opening date")+ylab("Density")+ggtitle("B)")+
    scale_fill_manual(values=c("#2c7bb6","#ffeda0","#d7191c"),
                      labels=c("0.05", "0.5", "1.0"))+
    scale_color_manual(values=c("#265a93","#ffdd87","#b71416"),
                      labels=c("0.05", "0.5", "1.0"))+
    theme(plot.title = element_text(hjust =-0.09,vjust=-6))+ylab(NULL)+
    scale_x_continuous(breaks=c(60,62,64,66,68,70,72)),
  ggplot(curves%>%filter(moment=="Skewness")%>%
           mutate(curve_id=factor(curve_id,levels=c("low","medium","high"))),
         aes(x=value,fill=curve_id,color=curve_id))+
    geom_density(alpha=0.5)+my_theme_legend()+theme(legend.position="top")+
    xlab("Opening date")+ylab("Density")+ggtitle("C)")+
    labs(fill="Skewness",color="Skewness")+
    scale_fill_manual(values=c("#2c7bb6","#ffeda0","#d7191c"),
                      labels=c("-1.1", "0.0", "1.1"))+
    scale_color_manual(values=c("#265a93","#ffdd87","#b71416"),
                      labels=c("-1.1", "0.0", "1.1"))+
    theme(plot.title = element_text(hjust =-0.09,vjust=-6))+ylab(NULL)+
    scale_x_continuous(breaks=c(62,64,66,68,70,72)),
  ggplot(curves%>%filter(moment=="Kurtosis")%>%
           mutate(curve_id=factor(curve_id,levels=c("low","medium","high"))),
         aes(x=value,fill=curve_id,color=curve_id))+
    geom_density(alpha=0.3)+my_theme_legend()+theme(legend.position="top")+
    labs(fill="Kurtosis",color="Kurtosis")+
    xlab("Opening date")+ylab("Density")+ggtitle("D)")+
    scale_fill_manual(values=c("#2c7bb6","#ffeda0","#d7191c"),
                      labels=c("2", "3", "4"))+
    scale_color_manual(values=c("#265a93","#ffdd87","#b71416"),
                      labels=c("2", "3", "4"))+
    theme(plot.title = element_text(hjust =-0.09,vjust=-6))+ylab(NULL)+
    scale_x_continuous(breaks=c(62,64,66,68,70,72)),
  ncol=4,
  left=textGrob("Density",
                gp=gpar(fontsize=14,fontfamily="serif"),rot = 90)
)

fig1B<-grid.arrange(
  # Mean
  ggplot(data_ids,aes(x=avFD_v))+
    geom_histogram(color="black",alpha=0.3,position="identity")+
    my_theme()+xlab("Mean")+ylab(NULL)+
    geom_vline(xintercept=60.1,color="#265a93",size=1.5,alpha=1)+
    geom_vline(xintercept=66.2,color="#ffdd87",size=1.5,alpha=1)+
    geom_vline(xintercept=72.4,color="#b71416",size=1.5,alpha=1)+
    theme(plot.title=element_text(hjust=-0.2,vjust=-4)),
  # Variance
  ggplot(data_ids,aes(x=var))+
    #geom_density(fill="grey")+
    geom_histogram(color="black",alpha=0.3,position="identity")+
    my_theme()+xlab("Variance")+ylab(NULL)+
    geom_vline(xintercept=0.05,color="#265a93",size=1.5,alpha=1)+
    geom_vline(xintercept=0.5,color="#ffdd87",size=1.5,alpha=1)+
    geom_vline(xintercept=1,color="#d7191c",size=1.5,alpha=1)+
    theme(plot.title=element_text(hjust=-0.2,vjust=-4)),
  # Skewness
  ggplot(data_ids,aes(x=skew))+
    #geom_density(fill="grey")+
    geom_histogram(color="black",alpha=0.3,position="identity")+
    my_theme()+xlab("Skewness")+ylab(NULL)+
    geom_vline(xintercept=-1.1,color="#265a93",size=1.5,alpha=1)+
    geom_vline(xintercept=0.0,color="#ffdd87",size=1.5,alpha=1)+
    geom_vline(xintercept=1.1,color="#b71416",size=1.5,alpha=1)+
    theme(plot.title=element_text(hjust=-0.2,vjust=-4)),
  # Kurtosis
  ggplot(data_ids,aes(x=kurt))+
    #geom_density(fill="grey")+
    geom_histogram(color="black",alpha=0.3,position="identity")+
    my_theme()+xlab("Kurtosis")+ylab(NULL)+
    geom_vline(xintercept=2,color="#265a93",size=1.5,alpha=1)+
    geom_vline(xintercept=3,color="#ffdd87",size=1.5,alpha=1)+
    geom_vline(xintercept=4,color="#b71416",size=1.5,alpha=1)+
    theme(plot.title=element_text(hjust=-0.2,vjust=-4)),
  ncol=4,widths=c(0.25,0.25,0.25,0.25),
  left=textGrob("Count",
                gp=gpar(fontsize=14,fontfamily="serif"),rot = 90))

fig1<-grid.arrange(fig1A,fig1B,nrow=2,
                       heights=c(0.55,0.45))
ggsave(file="output/figures/fig1.tiff",plot=fig1,width=15,height=8,
       dpi=300,compression="lzw")
```

# Q1: Within- vs. among-individual components of variation

## One model with all years

```{r eval=FALSE, include=FALSE}
brms_allyears<-brm(opening_date_v~1+(1|year)+(1|id),
                   data = data_id_flowers,
                   family = gaussian(),
                   chains = 4, cores = 4, iter = 10000, warmup = 3000, thin = 5,
                   control = list(adapt_delta = 0.98))
save(brms_allyears, file = "output/models/brms_allyears.rda")
```

```{r}
plot(brms_allyears)
```

```{r}
cowplot::plot_grid(mcmc_rhat_hist(brms::rhat(brms_allyears))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          mcmc_neff_hist(neff_ratio(brms_allyears))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          labels = c("A)","B)"),label_fontfamily="serif")
```

```{r}
summary(brms_allyears)
```

### Proportions of variance

```{r}
v_among_yrs <- (VarCorr(brms_allyears, summary = FALSE)$year$sd)^2
v_among_ids <- (VarCorr(brms_allyears, summary = FALSE)$id$sd)^2
v_within_ids <- (VarCorr(brms_allyears, summary = FALSE)$residual$sd)^2
prop_v_among_yrs <- as.mcmc(v_among_yrs / 
                              (v_among_yrs + v_among_ids + v_within_ids))
prop_v_among_ids <- as.mcmc(v_among_ids / 
                              (v_among_yrs + v_among_ids + v_within_ids))
prop_v_within_ids <- as.mcmc(v_within_ids / 
                              (v_among_yrs + v_among_ids + v_within_ids))
summary(prop_v_among_yrs)
summary(prop_v_among_ids)
summary(prop_v_within_ids)
plot(prop_v_among_yrs)
plot(prop_v_among_ids)
plot(prop_v_within_ids)
```

## Separate models for each year

```{r eval=FALSE, include=FALSE}
brms_87<-brm(opening_date_v~1+(1|id),
             data = subset(data_id_flowers,year==1987),
             family = gaussian(),
             chains = 4, cores = 4, iter = 10000, warmup = 3000, thin = 5)
save(brms_87, file = "output/models/brms_87.rda")
```

```{r eval=FALSE, include=FALSE}
brms_88<-brm(opening_date_v~1+(1|id),
             data = subset(data_id_flowers,year==1988),
             family = gaussian(),
             chains = 4, cores = 4, iter = 10000, warmup = 3000, thin = 5)
save(brms_88, file = "output/models/brms_88.rda")
```

```{r eval=FALSE, include=FALSE}
brms_89<-brm(opening_date_v~1+(1|id),
             data = subset(data_id_flowers,year==1989),
             family = gaussian(),
             chains = 4, cores = 4, iter = 10000, warmup = 3000, thin = 5)
save(brms_89, file = "output/models/brms_89.rda")
```

```{r}
plot(brms_87)
plot(brms_88)
plot(brms_89)
```

```{r}
cowplot::plot_grid(mcmc_rhat_hist(brms::rhat(brms_87))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          mcmc_neff_hist(neff_ratio(brms_87))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          labels = c("A)","B)"),label_fontfamily="serif")
cowplot::plot_grid(mcmc_rhat_hist(brms::rhat(brms_88))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          mcmc_neff_hist(neff_ratio(brms_88))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          labels = c("A)","B)"),label_fontfamily="serif")
cowplot::plot_grid(mcmc_rhat_hist(brms::rhat(brms_89))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          mcmc_neff_hist(neff_ratio(brms_89))+my_theme_legend()+
            theme(legend.position="top")+
            theme(legend.text=element_text(size=10)),
          labels = c("A)","B)"),label_fontfamily="serif")
```

```{r}
summary(brms_87)
summary(brms_88)
summary(brms_89)
```

### Proportions of variance

```{r}
v_among_87 <- (VarCorr(brms_87, summary = FALSE)$id$sd)^2
v_within_87 <- (VarCorr(brms_87, summary = FALSE)$residual$sd)^2
prop_v_among_87 <- as.mcmc(v_among_87 / (v_among_87 + v_within_87))
prop_v_within_87 <- as.mcmc(v_within_87 / (v_among_87 + v_within_87))
summary(prop_v_among_87)
summary(prop_v_within_87)
plot(prop_v_among_87)
plot(prop_v_within_87)
```

```{r}
v_among_88 <- (VarCorr(brms_88, summary = FALSE)$id$sd)^2
v_within_88 <- (VarCorr(brms_88, summary = FALSE)$residual$sd)^2
prop_v_among_88 <- as.mcmc(v_among_88 / (v_among_88 + v_within_88))
prop_v_within_88 <- as.mcmc(v_within_88 / (v_among_88 + v_within_88))
summary(prop_v_among_88)
summary(prop_v_within_88)
plot(prop_v_among_88)
plot(prop_v_within_88)
```

```{r}
v_among_89 <- (VarCorr(brms_89, summary = FALSE)$id$sd)^2
v_within_89 <- (VarCorr(brms_89, summary = FALSE)$residual$sd)^2
prop_v_among_89 <- as.mcmc(v_among_89 / (v_among_89 + v_within_89))
prop_v_within_89 <- as.mcmc(v_within_89 / (v_among_89 + v_within_89))
summary(prop_v_among_89)
summary(prop_v_within_89)
plot(prop_v_among_89)
plot(prop_v_within_89)
```

## Fgiure 2: Proportions of variance

```{r}
df_mean<-cbind(
  data.frame(among_years_mean=c(summary(prop_v_among_yrs)$statistics[1])),
  data.frame(among_ids_mean=c(summary(prop_v_among_ids)$statistics[1])),
  data.frame(within_ids_mean=c(summary(prop_v_within_ids)$statistics[1]))
  )
df_lower<-cbind(
  data.frame(among_years_lower=c(coda::HPDinterval(prop_v_among_yrs)[1])),
  data.frame(among_ids_lower=c(coda::HPDinterval(prop_v_among_ids)[1])),
  data.frame(within_ids_lower=c(coda::HPDinterval(prop_v_within_ids)[1]))
  )
rownames(df_lower)<-c("lower")
df_upper<-cbind(
  data.frame(among_years_upper=c(coda::HPDinterval(prop_v_among_yrs)[2])),
  data.frame(among_ids_upper=c(coda::HPDinterval(prop_v_among_ids)[2])),
  data.frame(within_ids_upper=c(coda::HPDinterval(prop_v_within_ids)[2]))
)
rownames(df_upper)<-c("upper")

fig2a<-full_join(
  full_join(
    df_mean%>%
      pivot_longer(cols=everything(),names_to="effect",values_to="mean")%>%
      mutate(effect=gsub("_mean", "", effect)),
    df_lower%>%
      pivot_longer(cols=everything(),names_to="effect",values_to="lower")%>%
      mutate(effect=gsub("_lower", "", effect))),
  df_upper%>%
    pivot_longer(cols=everything(),names_to="effect",values_to="upper")%>%
    mutate(effect=factor(gsub("_upper", "", effect))))%>%
  ggplot(aes(x=factor(effect,levels=c("among_years","among_ids","within_ids")),
             y=mean,ymin=lower,ymax=upper,color=effect))+
  geom_errorbar(linewidth=0.5,width=0.2,position=position_dodge(0.3))+
  geom_point(size=2,position=position_dodge(0.3))+
  my_theme()+xlab(NULL)+ylab(NULL)+
  scale_x_discrete(labels=c("among_ids"="Among\nindividuals",
                            "among_years"="Among\nyears",
                            "within_ids"="Within\nindividuals"))+
  scale_y_continuous(limits=c(0,1),breaks=c(0,0.25,0.50,0.75,1))+
  scale_color_manual(values=c("#F8766D","black","#00BFC4"))+ggtitle("A)")+
  theme(plot.title=element_text(face="plain",size=16,hjust=-0.15,vjust=-0.5))
fig2b<-full_join(
  data.frame(among_mean=c(summary(prop_v_among_87)$statistics[1],
                     summary(prop_v_among_88)$statistics[1],
                     summary(prop_v_among_89)$statistics[1]))%>%
    mutate(year=c(1987,1988,1989)),
  data.frame(within_mean=c(summary(prop_v_within_87)$statistics[1],
                      summary(prop_v_within_88)$statistics[1],
                      summary(prop_v_within_89)$statistics[1]))%>%
    mutate(year=c(1987,1988,1989)))%>%
    pivot_longer(cols=c("among_mean","within_mean"),
                 names_to="effect",values_to="mean")%>%
  mutate(effect=ifelse(effect=="among_mean","among","within"))%>%
  full_join(data.frame(among_lower=c(coda::HPDinterval(prop_v_among_87)[1],
                                     coda::HPDinterval(prop_v_among_88)[1],
                                     coda::HPDinterval(prop_v_among_89)[1]),
                       among_upper=c(coda::HPDinterval(prop_v_among_87)[2],
                                     coda::HPDinterval(prop_v_among_88)[2],
                                     coda::HPDinterval(prop_v_among_89)[2]),
                       within_lower=c(coda::HPDinterval(prop_v_within_87)[1],
                                      coda::HPDinterval(prop_v_within_88)[1],
                                      coda::HPDinterval(prop_v_within_89)[1]),
                       within_upper=c(coda::HPDinterval(prop_v_within_87)[2],
                                      coda::HPDinterval(prop_v_within_88)[2],
                                      coda::HPDinterval(prop_v_within_89)[2]))%>%
              mutate(year=c(1987,1988,1989))%>%
            pivot_longer(cols=c("among_lower","among_upper",
                                "within_lower","within_upper"),
                 names_to=c("effect","cat"),names_sep="_",values_to="value")%>%
              pivot_wider(names_from="cat",values_from="value"))%>%
  mutate(year=factor(year))%>%
  ggplot(aes(x=year,y=mean,ymin=lower,ymax=upper,color=effect))+
  geom_errorbar(linewidth=0.5,width=0.2,position=position_dodge(0.3))+
  geom_point(size=2,position=position_dodge(0.3))+
  my_theme()+xlab("Year")+ylab(NULL)+ggtitle("B)")+
  scale_y_continuous(limits=c(0,1),breaks=c(0,0.25,0.50,0.75,1))+
  theme(plot.title=element_text(face="plain",size=16,hjust=-0.15,vjust=-0.5))
fig2<-grid.arrange(fig2a,fig2b,ncol=2,
                   left=textGrob("Proportion of variance",
                                 gp=gpar(fontsize=14,fontfamily="serif"),
                                 rot = 90))
ggsave(file="output/figures/fig2.tiff",plot=fig2,width=8, height=3.5,
       dpi=300,compression="lzw")
```

# Figure 3: Differences in moments and fitness components among years

```{r}
Anova(lm(avFD_v~year,data_ids))
Anova(lm(var~year,data_ids))
Anova(lm(skew~year,data_ids))
Anova(lm(kurt~year,data_ids))
```

```{r}
letters_mean<-as.data.frame.list(
  multcompLetters4(aov(avFD_v~year,data_ids),
                   TukeyHSD(aov(avFD_v~year,data_ids)))$year)%>%
  dplyr::select(Letters)%>%
  rownames_to_column(var="year")
letters_var<-as.data.frame.list(
  multcompLetters4(aov(var~year,data_ids),
                   TukeyHSD(aov(var~year,data_ids)))$year)%>%
  dplyr::select(Letters)%>%
  rownames_to_column(var="year")
letters_skew<-as.data.frame.list(
  multcompLetters4(aov(skew~year,data_ids),
                   TukeyHSD(aov(skew~year,data_ids)))$year)%>%
  dplyr::select(Letters)%>%
  rownames_to_column(var="year")
letters_kurt<-as.data.frame.list(
  multcompLetters4(aov(kurt~year,data_ids),
                   TukeyHSD(aov(kurt~year,data_ids)))$year)%>%
  dplyr::select(Letters)%>%
  rownames_to_column(var="year")
```

```{r}
# Fruit set
Anova(glm(cbind(fr_set=n_mat_intact_fr,
                fr_no_set=n_fl-n_mat_intact_fr)~year,data_ids,
          family="binomial")) 
summary(glht(glm(cbind(fr_set=n_mat_intact_fr,fr_no_set=n_fl-n_mat_intact_fr)~
                   year,data_ids,family="binomial"),
             linfct=mcp(year="Tukey")))
# Proportion of seeds escaping predation
Anova(glm(cbind(npr_seeds=round(n_seed)-round(n_preyed_seed),
                pr_seeds=round(n_preyed_seed))~year,
          data_ids,family="binomial"))
summary(glht(glm(cbind(npr_seeds=round(n_seed)-round(n_preyed_seed),
                       pr_seeds=round(n_preyed_seed))~year,data_ids,
                 family="binomial"),
             linfct=mcp(year="Tukey")))
# Fitness
Anova(lm(fitness~year,data_ids))
summary(glht(lm(fitness~year,data_ids),linfct=mcp(year="Tukey")))
```

```{r}
letters_fr_set<-data.frame(
  Letters=cld(glht(glm(cbind(fr_set=n_mat_intact_fr,
                             fr_no_set=n_fl-n_mat_intact_fr)~
                   year,data_ids,family="binomial"),
                   linfc= mcp(year="Tukey")))$mcletters$Letters)%>%
  rownames_to_column(var="year")
letters_seeds_escpred<-data.frame(
  Letters=cld(glht(glm(cbind(npr_seeds=round(n_seed)-round(n_preyed_seed),
                       pr_seeds=round(n_preyed_seed))~year,data_ids,
                 family="binomial"),
                   linfc= mcp(year="Tukey")))$mcletters$Letters)%>%
  rownames_to_column(var="year")
letters_sel<-data.frame(
  Letters=cld(glht(lm(fitness~year,data_ids),
                   linfc= mcp(year="Tukey")))$mcletters$Letters)%>%
  rownames_to_column(var="year")

```

```{r}
data_ids%>%group_by(year)%>%
  summarise(mean_fr_set=mean(fr_set),
            mean_prop_seed_preyed=mean(prop_seed_preyed,na.rm=T),
            mean_fitness=mean(fitness))
```

```{r}
data_ids$prop_seeds_escpred<-1-data_ids$prop_seed_preyed
```

```{r}
fig3<-ggarrange(
  ggerrorplot(data_ids,x="year",y="avFD_v",combine=F,size=0.5,
              desc_stat="mean_se",error.plot = "errorbar",add = "none",
              xlab=NULL)+my_theme()+ylab("Mean")+xlab(NULL)+
    stat_summary(geom="point",size=2,fun.y="mean")+
    geom_text(data=data_ids%>%group_by(year)%>%
                summarize(mean_se=mean(avFD_v)+(sd(avFD_v)/sqrt(n()))),
              aes(x=year,y=mean_se+0.1,label=letters_mean$Letters),
              vjust=-0.2,family="serif")+
    ggtitle("A)")+  
    theme(plot.title=element_text(face="plain",size=16,hjust=-0.5,vjust=-5)),
  ggerrorplot(data_ids,x="year",y="var",combine=F,size=0.5,
              desc_stat="mean_se",error.plot = "errorbar",add = "none",
              xlab=NULL)+my_theme()+ylab("Variance")+xlab(NULL)+
    stat_summary(geom="point",size=2,fun.y="mean")+
    geom_text(data=data_ids%>%group_by(year)%>%
                summarize(mean_se=mean(var,na.rm=T)+
                            (sd(var,na.rm=T)/sqrt(n()))),
              aes(x=year,y=mean_se+0.005,label=letters_var$Letters),
              vjust=-0.2,family="serif")+
    ggtitle("B)")+
    theme(plot.title=element_text(face="plain",size=16,hjust=-0.5,vjust=-5)),
  ggerrorplot(data_ids,x="year",y="skew",combine=F,size=0.5,
              desc_stat="mean_se",error.plot = "errorbar",add = "none",
              xlab=NULL)+my_theme()+ylab("Skewness")+xlab(NULL)+
    stat_summary(geom="point",size=2,fun.y="mean")+
    geom_text(data=data_ids%>%group_by(year)%>%
                summarize(mean_se=mean(skew,na.rm=T)+
                            (sd(skew,na.rm=T)/sqrt(n()))),
              aes(x=year,y=mean_se+0.007,label=letters_skew$Letters),
              vjust=-0.2,family="serif")+
    ggtitle("C)")+
    theme(plot.title=element_text(face="plain",size=16,hjust=-0.5,vjust=-5)),
  ggerrorplot(data_ids,x="year",y="kurt",combine=F,size=0.5,
              desc_stat="mean_se",error.plot = "errorbar",add = "none",
              xlab=NULL)+my_theme()+ylab("Kurtosis")+xlab(NULL)+
    stat_summary(geom="point",size=2,fun.y="mean")+
    geom_text(data=data_ids%>%group_by(year)%>%
                summarize(mean_se=mean(kurt,na.rm=T)+
                            (sd(kurt,na.rm=T)/sqrt(n()))),
              aes(x=year,y=mean_se+0.009,label=letters_kurt$Letters),
              vjust=-0.2,family="serif")+
    ggtitle("D)")+
    theme(plot.title=element_text(face="plain",size=16,hjust=-0.5,vjust=-5)),
  ggerrorplot(data_ids,x="year",y="fr_set",combine=F,size=0.5,
              desc_stat="mean_se",error.plot = "errorbar",add = "none",
              xlab=NULL)+my_theme()+ylab("Fruit set")+xlab(NULL)+
    stat_summary(geom="point",size=2,fun.y="mean")+
    geom_text(data=data_ids%>%group_by(year)%>%
                summarize(mean_se=mean(fr_set)+(sd(fr_set)/sqrt(n()))),
              aes(x=year,y=mean_se+0.001,label=letters_fr_set$Letters),
              vjust=-0.2,family="serif")+
    ggtitle("E)")+
    theme(plot.title=element_text(face="plain",size=16,hjust=-0.5,vjust=-5)),
  ggerrorplot(data_ids,x="year",y="prop_seeds_escpred",combine=F,size=0.5,
              desc_stat="mean_se",error.plot = "errorbar",add = "none",
              xlab=NULL)+my_theme()+
    ylab("Propotion of seeds\nescaping predation")+xlab(NULL)+
    stat_summary(geom="point",size=2,fun.y="mean")+
    geom_text(data=data_ids%>%group_by(year)%>%
                summarize(mean_se=mean(prop_seeds_escpred,na.rm=T)+
                            (sd(prop_seeds_escpred,na.rm=T)/sqrt(n()))),
              aes(x=year,y=mean_se+0.008,label=letters_seeds_escpred$Letters),
              vjust=-0.2,family="serif")+
    ggtitle("F)")+
    theme(plot.title=element_text(face="plain",size=16,hjust=-0.5,vjust=-5)),
  ggerrorplot(data_ids,x="year",y="fitness",combine=F,size=0.5,
              desc_stat="mean_se",error.plot = "errorbar",add = "none",
              xlab=NULL)+my_theme()+ylab("Fitness")+xlab(NULL)+
    stat_summary(geom="point",size=2,fun.y="mean")+
    geom_text(data=data_ids%>%group_by(year)%>%
                summarize(mean_se=mean(fitness)+(sd(fitness)/sqrt(n()))),
              aes(x=year,y=mean_se+0.008,label=letters_sel$Letters),
              vjust=-0.2,family="serif")+
    ggtitle("G)")+
    theme(plot.title=element_text(face="plain",size=16,hjust=-0.5,vjust=-5)),
  nrow=2,ncol=4,align="v")
fig3
ggsave(fig3,
       filename="output/figures/fig3.tiff",
       device="tiff",width=30,height=16,units="cm",dpi=300,compression="lzw")
```

# Appendix 4: Correlations of the 4 moments using all data

```{r}
data_ids%>%dplyr::select(year,avFD_v,var,skew,kurt) %>%
  rename(Mean=avFD_v,Variance=var,Skewness=skew,Kurtosis=kurt)%>%
  ungroup()%>%
  select_if(is.numeric) %>% 
  cor(use="pairwise.complete.obs") %>%
  ggcorrplot(method = "square", type = "lower", hc.order = F, lab = T,
             legend.title="Correlation")+
  theme(legend.position="right")+my_theme_legend()+xlab(NULL)+ylab(NULL)
ggsave(filename="output/figures/Appendix4.tiff",
       device="tiff",width=12,height=8,units="cm",dpi=300,compression="lzw")
```

# Appendix 5: Correlations of the 4 moments for each year

```{r}
Appendix5_legend<-get_legend(
  data_ids%>%dplyr::select(year,avFD_v,var,skew,kurt) %>%
    filter(year==1987) %>%
    rename(Mean=avFD_v,Variance=var,Skewness=skew,Kurtosis=kurt)%>%
    select_if(is.numeric) %>% 
    cor(use="pairwise.complete.obs") %>%
    ggcorrplot(method = "square", type = "lower", hc.order = F, lab = T,
             legend.title="Correlation")+
    ggtitle("1987")+my_theme_legend()+theme(legend.position="right")+
    xlab(NULL)+ylab(NULL)
)
Appendix5<-grid.arrange(
  data_ids%>%dplyr::select(year,avFD_v,var,skew,kurt) %>%
    filter(year==1987) %>%
    rename(Mean=avFD_v,Variance=var,Skewness=skew,Kurtosis=kurt)%>%
    select_if(is.numeric) %>% 
    cor(use="pairwise.complete.obs") %>%
    ggcorrplot(method = "square", type = "lower", hc.order = F, lab = T,
             legend.title="Correlation")+
    ggtitle("1987")+theme(legend.position="none")+
    my_theme()+xlab(NULL)+ylab(NULL),
  data_ids%>%dplyr::select(year,avFD_v,var,skew,kurt) %>%
    filter(year==1988) %>%
    rename(Mean=avFD_v,Variance=var,Skewness=skew,Kurtosis=kurt)%>%
    select_if(is.numeric) %>% 
    cor(use="pairwise.complete.obs") %>%
    ggcorrplot(method = "square", type = "lower", hc.order = F, lab = T,
             legend.title="Correlation")+
    ggtitle("1988")+theme(legend.position="none")+
    my_theme()+xlab(NULL)+ylab(NULL),
  data_ids%>%dplyr::select(year,avFD_v,var,skew,kurt) %>%
    filter(year==1989) %>%
    rename(Mean=avFD_v,Variance=var,Skewness=skew,Kurtosis=kurt)%>%
    select_if(is.numeric) %>% 
    cor(use="pairwise.complete.obs") %>%
    ggcorrplot(method = "square", type = "lower", hc.order = F, lab = T,
             legend.title="Correlation")+
    ggtitle("1989")+theme(legend.position="none")+
    my_theme()+xlab(NULL)+ylab(NULL),
  Appendix5_legend,
  ncol=4,widths=c(0.3,0.3,0.3,0.1))
ggsave(Appendix5,filename="output/figures/Appendix5.tiff",
       device="tiff",width=30,height=10,units="cm",dpi=300,compression="lzw")
```

# Q2: Phenotypic selection on within-individual variation

Phenotypic selection models with n_fl

n_fl --> log --> std

```{r}
data_ids<-data_ids%>%
  mutate(n_fl_log=log(n_fl))%>%
  group_by(year)%>%
  mutate(n_fl_log_std=as.vector(scale(n_fl_log)))
```

Using (within-year) standardized traits.

## Appendix 6: Global models phenotypic selection with n_fl

Including only linear effects and interactions with year. Mixed models with id as a random effect.

```{r}
sel_nfl_linear<-lmer(fitness_rel~avFD_std+var_std+skew_std+kurt_std+n_fl_log_std+
              avFD_std:year+var_std:year+skew_std:year+kurt_std:year+
                (1|id),
              # Not including main effect of yr cause fitness rel within yrs
              subset(data_ids,!is.na(avFD_std)&!is.na(var_std)&
                       !is.na(skew_std)&!is.na(kurt_std)))
tab_model(sel_nfl_linear,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("Selection, linear effects"))
Anova(sel_nfl_linear)
check_collinearity(lm(fitness_rel~avFD_std+var_std+skew_std+kurt_std+n_fl_std,
              subset(data_ids,!is.na(avFD_std)&!is.na(var_std)&
                       !is.na(skew_std)&!is.na(kurt_std))))
```

According to Anova, no differences among years.

### Permutational ANOVA

```{r eval=FALSE, include=FALSE}
permANOVA_sel_nfl_linear<-permmodels(sel_nfl_linear, nperm=10000,type="II",
                                     ncore=16,seed=123)
save(permANOVA_sel_nfl_linear,file="output/permANOVA_sel_nfl_linear.RData")
```

```{r}
permANOVA_sel_nfl_linear$ANOVA
```

### Bootstraped confidence intervals

```{r eval=FALSE, include=FALSE}
b_sel_nfl_linear<-bootMer(x=sel_nfl_linear,FUN=fixef,nsim=10000)
save(b_sel_nfl_linear,file="output/b_sel_nfl_linear.RData")
```

```{r}
confint(b_sel_nfl_linear,level=0.95, method="boot") # Percentile method
```

### Contrasts for all years

```{r}
emtrends(sel_nfl_linear,pairwise~year, var = "avFD_std",adjust="none")$contrasts
emtrends(sel_nfl_linear,pairwise~year, var = "var_std",adjust="none")$contrasts
emtrends(sel_nfl_linear,pairwise~year, var = "skew_std",adjust="none")$contrasts
emtrends(sel_nfl_linear,pairwise~year, var = "kurt_std",adjust="none")$contrasts
```

## Table 1: Yearly models

```{r}
sel_nfl_linear_87<-lm(fitness_rel~avFD_std+var_std+skew_std+kurt_std+
                        n_fl_log_std,
                      subset(data_ids,year==1987&
                               !is.na(avFD_std)&!is.na(var_std)&
                               !is.na(skew_std)&!is.na(kurt_std)))
sel_nfl_linear_88<-lm(fitness_rel~avFD_std+var_std+skew_std+kurt_std+
                        n_fl_log_std,
                      subset(data_ids,year==1988&
                               !is.na(avFD_std)&!is.na(var_std)&
                               !is.na(skew_std)&!is.na(kurt_std)))
sel_nfl_linear_89<-lm(fitness_rel~avFD_std+var_std+skew_std+kurt_std+
                        n_fl_log_std,
                      subset(data_ids,year==1989&
                               !is.na(avFD_std)&!is.na(var_std)&
                               !is.na(skew_std)&!is.na(kurt_std)))
tab_model(sel_nfl_linear_87,sel_nfl_linear_88,sel_nfl_linear_89,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("1987","1988","1989"),
          title="Selection, linear effects")
check_collinearity(sel_nfl_linear_87)
check_collinearity(sel_nfl_linear_88)
check_collinearity(sel_nfl_linear_89)
Anova(sel_nfl_linear_87)
Anova(sel_nfl_linear_88)
Anova(sel_nfl_linear_89)
```

### BCA intervals

#### 1987

```{r eval=FALSE, include=FALSE}
# avFD_std
slp <- function(sel_nfl_linear_87) coef(sel_nfl_linear_87)[2]
b <- car::Boot(sel_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
avFD_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# var_std
slp <- function(sel_nfl_linear_87) coef(sel_nfl_linear_87)[3]
b <- car::Boot(sel_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
var_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# skew_std
slp <- function(sel_nfl_linear_87) coef(sel_nfl_linear_87)[4]
b <- car::Boot(sel_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
skew_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# kurt_std
slp <- function(sel_nfl_linear_87) coef(sel_nfl_linear_87)[5]
b <- car::Boot(sel_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
kurt_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# n_fl_log_std
slp <- function(sel_nfl_linear_87) coef(sel_nfl_linear_87)[6]
b <- car::Boot(sel_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
n_fl_log_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_sel_nfl_linear_87 <- cbind(
  rbind(avFD_std_ci[1,],var_std_ci[1,], skew_std_ci[1,], kurt_std_ci[1,],
        n_fl_log_std_ci[1,]),
  rbind(avFD_std_ci[2,],var_std_ci[2,], skew_std_ci[2,], kurt_std_ci[2,],
        n_fl_log_std_ci[2,])
)
colnames(BCIs_sel_nfl_linear_87)<-c("lower","upper")
rownames(BCIs_sel_nfl_linear_87)<-c("avFD_std","var_std","skew_std","kurt_std",
                                    "n_fl_log_std")
save(BCIs_sel_nfl_linear_87,file="output/BCIs_sel_nfl_linear_87.RData")
```

```{r}
BCIs_sel_nfl_linear_87
```

#### 1988

```{r eval=FALSE, include=FALSE}
# avFD_std
slp <- function(sel_nfl_linear_88) coef(sel_nfl_linear_88)[2]
b <- car::Boot(sel_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
avFD_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# var_std
slp <- function(sel_nfl_linear_88) coef(sel_nfl_linear_88)[3]
b <- car::Boot(sel_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
var_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# skew_std
slp <- function(sel_nfl_linear_88) coef(sel_nfl_linear_88)[4]
b <- car::Boot(sel_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
skew_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# kurt_std
slp <- function(sel_nfl_linear_88) coef(sel_nfl_linear_88)[5]
b <- car::Boot(sel_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
kurt_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# n_fl_log_std
slp <- function(sel_nfl_linear_88) coef(sel_nfl_linear_88)[6]
b <- car::Boot(sel_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
n_fl_log_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_sel_nfl_linear_88 <- cbind(
  rbind(avFD_std_ci[1,],var_std_ci[1,], skew_std_ci[1,], kurt_std_ci[1,],
        n_fl_log_std_ci[1,]),
  rbind(avFD_std_ci[2,],var_std_ci[2,], skew_std_ci[2,], kurt_std_ci[2,],
        n_fl_log_std_ci[2,])
)
colnames(BCIs_sel_nfl_linear_88)<-c("lower","upper")
rownames(BCIs_sel_nfl_linear_88)<-c("avFD_std","var_std","skew_std","kurt_std",
                                    "n_fl_log_std")
save(BCIs_sel_nfl_linear_88,file="output/BCIs_sel_nfl_linear_88.RData")
```

```{r}
BCIs_sel_nfl_linear_88
```

#### 1989

```{r eval=FALSE, include=FALSE}
# avFD_std
slp <- function(sel_nfl_linear_89) coef(sel_nfl_linear_89)[2]
b <- car::Boot(sel_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
avFD_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# var_std
slp <- function(sel_nfl_linear_89) coef(sel_nfl_linear_89)[3]
b <- car::Boot(sel_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
var_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# skew_std
slp <- function(sel_nfl_linear_89) coef(sel_nfl_linear_89)[4]
b <- car::Boot(sel_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
skew_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# kurt_std
slp <- function(sel_nfl_linear_89) coef(sel_nfl_linear_89)[5]
b <- car::Boot(sel_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
kurt_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# n_fl_log_std
slp <- function(sel_nfl_linear_89) coef(sel_nfl_linear_89)[6]
b <- car::Boot(sel_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
n_fl_log_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_sel_nfl_linear_89 <- cbind(
  rbind(avFD_std_ci[1,],var_std_ci[1,], skew_std_ci[1,], kurt_std_ci[1,],
        n_fl_log_std_ci[1,]),
  rbind(avFD_std_ci[2,],var_std_ci[2,], skew_std_ci[2,], kurt_std_ci[2,],
        n_fl_log_std_ci[2,])
)
colnames(BCIs_sel_nfl_linear_89)<-c("lower","upper")
rownames(BCIs_sel_nfl_linear_89)<-c("avFD_std","var_std","skew_std","kurt_std",
                                    "n_fl_log_std")
save(BCIs_sel_nfl_linear_89,file="output/BCIs_sel_nfl_linear_89.RData")
```

```{r}
BCIs_sel_nfl_linear_89
```

## Figure 4: Effects on relative iftness

```{r}
coefs_sel_nfl_linear<-rbind(
  tidy(sel_nfl_linear_87)[2:5,]%>%mutate(year=as.factor("1987")),
  tidy(sel_nfl_linear_88)[2:5,]%>%mutate(year=as.factor("1988")),
  tidy(sel_nfl_linear_89)[2:5,]%>%mutate(year=as.factor("1989"))
  )%>%
  mutate(signif=ifelse(p.value>0.05,0,1),
         coef=format(round(estimate,3),nsmall=3),
         star=ifelse(signif==1,"*","  "),
         print=paste(coef,star),
         predictor=term,
         type=as.factor("linear"))%>%
  mutate(signif=factor(ifelse(row_number()==3,1,signif)))
# Change signif for skewness cause it is signif according to BCA intervals
```

```{r}
prediction_sel_nfl<-rbind(rbind(tibble(ggpredict(sel_nfl_linear_87,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(sel_nfl_linear_88,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(sel_nfl_linear_89,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="avFD_std"), # avFD_std
                    rbind(tibble(ggpredict(sel_nfl_linear_87,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(sel_nfl_linear_88,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(sel_nfl_linear_89,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="var_std"), # var_std
                    rbind(tibble(ggpredict(sel_nfl_linear_87,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(sel_nfl_linear_88,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(sel_nfl_linear_89,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="skew_std"), # skew_std
                    rbind(tibble(ggpredict(sel_nfl_linear_87,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(sel_nfl_linear_88,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(sel_nfl_linear_89,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="kurt_std"))%>% # kurt_std
  mutate(response="sel_nfl")%>%
  left_join(coefs_sel_nfl_linear%>%
              dplyr::select(predictor,coef,year,signif,star,print,type))%>%
  mutate(predictor=factor(predictor,
                          levels=c("avFD_std","var_std","skew_std","kurt_std")))
```

```{r}
predictor.labs<-c("Mean","Variance","Skewness","Kurtosis")
names(predictor.labs)<-c("avFD_std","var_std","skew_std","kurt_std")
```

```{r}
fig4<-grid.arrange(
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_sel_nfl%>%
                  filter(predictor=="avFD_std"), 
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_sel_nfl%>%
                filter(predictor=="avFD_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fitness_rel,avFD_std)%>%
                 pivot_longer(cols=avFD_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("avFD_std"))),
               aes(x=value_predictor,y=fitness_rel),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dashed", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_sel_nfl%>%
                  filter(predictor=="var_std"), 
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_sel_nfl%>%
                filter(predictor=="var_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fitness_rel,var_std)%>%
                 pivot_longer(cols=var_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("var_std"))),
               aes(x=value_predictor,y=fitness_rel),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dashed", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_sel_nfl%>%
                  filter(predictor=="skew_std"), 
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_sel_nfl%>%
                filter(predictor=="skew_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fitness_rel,skew_std)%>%
                 pivot_longer(cols=skew_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("skew_std"))),
               aes(x=value_predictor,y=fitness_rel),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dashed", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_sel_nfl%>%
                  filter(predictor=="kurt_std"), 
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_sel_nfl%>%
                filter(predictor=="kurt_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fitness_rel,kurt_std)%>%
                 pivot_longer(cols=kurt_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("kurt_std"))),
               aes(x=value_predictor,y=fitness_rel),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dashed", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ncol=1,
  left=textGrob("Relative fitness (number of intact seeds)",
                gp=gpar(fontsize=14,fontfamily="serif"),rot = 90),
  bottom=textGrob("Value of the moment",
                  gp=gpar(fontsize=14,fontfamily="serif")),
  top=textGrob("       1987                                      1988                                        1989",
               gp=gpar(fontsize=14,fontfamily="serif"))
)
ggsave(file="output/figures/fig4.tiff", plot=fig4, width=8, height=8,dpi=300,
       compression="lzw")
```

## Appendix 7: Global models ABSOLUTE FITNESS ~ ABSOLUTE TRAITS

Including only linear effects and interactions with year.

```{r}
abs_nfl_linear<-lmer(fitness~avFD+var+skew+kurt+n_fl_log+year+
              avFD:year+var:year+skew:year+kurt:year+
              # Including main effect of yr cause fitness is not rel within yrs
                (1|id),
              subset(data_ids,!is.na(avFD_std)&!is.na(var_std)&
                       !is.na(skew_std)&!is.na(kurt_std)))
tab_model(abs_nfl_linear,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("Absolute fitness and traits, linear effects"))
Anova(abs_nfl_linear)
check_collinearity(lm(fitness~avFD+var+skew+kurt+n_fl_log,
              subset(data_ids,!is.na(avFD_std)&!is.na(var_std)&
                       !is.na(skew_std)&!is.na(kurt_std))))
```

### Permutational ANOVA

```{r eval=FALSE, include=FALSE}
permANOVA_abs_nfl_linear<-permmodels(abs_nfl_linear, nperm=10000,type="II",
                                     ncore=16,seed=123)
save(permANOVA_abs_nfl_linear,file="output/permANOVA_abs_nfl_linear.RData")
```

```{r}
permANOVA_abs_nfl_linear$ANOVA
```

### Bootstraped confidence intervals

```{r eval=FALSE, include=FALSE}
b_abs_nfl_linear<-bootMer(x=abs_nfl_linear,FUN=fixef,nsim=10000)
save(b_abs_nfl_linear,file="output/b_abs_nfl_linear.RData")
```

```{r}
confint(b_abs_nfl_linear,level=0.95, method="boot") # Percentile method
```

### Contrasts for all years

```{r}
emtrends(abs_nfl_linear,pairwise~year, var = "avFD",adjust="none")$contrasts
emtrends(abs_nfl_linear,pairwise~year, var = "var",adjust="none")$contrasts
emtrends(abs_nfl_linear,pairwise~year, var = "skew",adjust="none")$contrasts
emtrends(abs_nfl_linear,pairwise~year, var = "kurt",adjust="none")$contrasts
```

## Appendix 8: Yearly models ABSOLUTE FITNESS ~ ABSOLUTE TRAITS

```{r}
abs_nfl_linear_87<-lm(fitness~avFD+var+skew+kurt+
                        n_fl_log,
                      subset(data_ids,year==1987&
                               !is.na(avFD)&!is.na(var)&
                               !is.na(skew)&!is.na(kurt)))
abs_nfl_linear_88<-lm(fitness_rel~avFD+var+skew+kurt+
                        n_fl_log,
                      subset(data_ids,year==1988&
                               !is.na(avFD)&!is.na(var)&
                               !is.na(skew)&!is.na(kurt)))
abs_nfl_linear_89<-lm(fitness_rel~avFD+var+skew+kurt+
                        n_fl_log,
                      subset(data_ids,year==1989&
                               !is.na(avFD)&!is.na(var)&
                               !is.na(skew)&!is.na(kurt)))
tab_model(abs_nfl_linear_87,abs_nfl_linear_88,abs_nfl_linear_89,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("1987","1988","1989"),
          title="Absolute fitness and traits, linear effects")
```

#### BCA intervals

##### 1987

```{r eval=FALSE, include=FALSE}
# avFD
slp <- function(abs_nfl_linear_87) coef(abs_nfl_linear_87)[2]
b <- car::Boot(abs_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
avFD_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# var
slp <- function(abs_nfl_linear_87) coef(abs_nfl_linear_87)[3]
b <- car::Boot(abs_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
var_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# skew
slp <- function(abs_nfl_linear_87) coef(abs_nfl_linear_87)[4]
b <- car::Boot(abs_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
skew_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# kurt
slp <- function(abs_nfl_linear_87) coef(abs_nfl_linear_87)[5]
b <- car::Boot(abs_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
kurt_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# n_fl_log
slp <- function(abs_nfl_linear_87) coef(abs_nfl_linear_87)[6]
b <- car::Boot(abs_nfl_linear_87,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
n_fl_log_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_abs_nfl_linear_87 <- cbind(
  rbind(avFD_ci[1,],var_ci[1,], skew_ci[1,], kurt_ci[1,],
        n_fl_log_ci[1,]),
  rbind(avFD_ci[2,],var_ci[2,], skew_ci[2,], kurt_ci[2,],
        n_fl_log_ci[2,])
)
colnames(BCIs_abs_nfl_linear_87)<-c("lower","upper")
rownames(BCIs_abs_nfl_linear_87)<-c("avFD","var","skew","kurt",
                                    "n_fl_log")
save(BCIs_abs_nfl_linear_87,file="output/BCIs_abs_nfl_linear_87.RData")
```

```{r}
BCIs_abs_nfl_linear_87
```

##### 1988

```{r eval=FALSE, include=FALSE}
# avFD
slp <- function(abs_nfl_linear_88) coef(abs_nfl_linear_88)[2]
b <- car::Boot(abs_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
avFD_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# var
slp <- function(abs_nfl_linear_88) coef(abs_nfl_linear_88)[3]
b <- car::Boot(abs_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
var_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# skew
slp <- function(abs_nfl_linear_88) coef(abs_nfl_linear_88)[4]
b <- car::Boot(abs_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
skew_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# kurt
slp <- function(abs_nfl_linear_88) coef(abs_nfl_linear_88)[5]
b <- car::Boot(abs_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
kurt_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# n_fl_log
slp <- function(abs_nfl_linear_88) coef(abs_nfl_linear_88)[6]
b <- car::Boot(abs_nfl_linear_88,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
n_fl_log_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_abs_nfl_linear_88 <- cbind(
  rbind(avFD_ci[1,],var_ci[1,], skew_ci[1,], kurt_ci[1,],
        n_fl_log_ci[1,]),
  rbind(avFD_ci[2,],var_ci[2,], skew_ci[2,], kurt_ci[2,],
        n_fl_log_ci[2,])
)
colnames(BCIs_abs_nfl_linear_88)<-c("lower","upper")
rownames(BCIs_abs_nfl_linear_88)<-c("avFD","var","skew","kurt",
                                    "n_fl_log")
save(BCIs_abs_nfl_linear_88,file="output/BCIs_abs_nfl_linear_88.RData")
```

```{r}
BCIs_abs_nfl_linear_88
```

##### 1989

```{r eval=FALSE, include=FALSE}
# avFD
slp <- function(abs_nfl_linear_89) coef(abs_nfl_linear_89)[2]
b <- car::Boot(abs_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
avFD_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# var
slp <- function(abs_nfl_linear_89) coef(abs_nfl_linear_89)[3]
b <- car::Boot(abs_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
var_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# skew
slp <- function(abs_nfl_linear_89) coef(abs_nfl_linear_89)[4]
b <- car::Boot(abs_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
skew_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# kurt
slp <- function(abs_nfl_linear_89) coef(abs_nfl_linear_89)[5]
b <- car::Boot(abs_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
kurt_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# n_fl_log
slp <- function(abs_nfl_linear_89) coef(abs_nfl_linear_89)[6]
b <- car::Boot(abs_nfl_linear_89,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
n_fl_log_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_abs_nfl_linear_89 <- cbind(
  rbind(avFD_ci[1,],var_ci[1,], skew_ci[1,], kurt_ci[1,],
        n_fl_log_ci[1,]),
  rbind(avFD_ci[2,],var_ci[2,], skew_ci[2,], kurt_ci[2,],
        n_fl_log_ci[2,])
)
colnames(BCIs_abs_nfl_linear_89)<-c("lower","upper")
rownames(BCIs_abs_nfl_linear_89)<-c("avFD","var","skew","kurt",
                                    "n_fl_log")
save(BCIs_abs_nfl_linear_89,file="output/BCIs_abs_nfl_linear_89.RData")
```

```{r}
BCIs_abs_nfl_linear_89
```

## Appendix 9: Global models without n_fl

Including only linear effects and interactions with year. Mixed models with id as a random effect.

```{r}
sel_linear<-lmer(fitness_rel~avFD_std+var_std+skew_std+kurt_std+
              avFD_std:year+var_std:year+skew_std:year+kurt_std:year+
              # Not including main effect of yr cause fitness rel within yrs
                (1|id),
              subset(data_ids,!is.na(avFD_std)&!is.na(var_std)&
                       !is.na(skew_std)&!is.na(kurt_std)))
tab_model(sel_linear,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("Selection, linear effects, without n_fl"))
Anova(sel_linear)
check_collinearity(lm(fitness_rel~avFD_std+var_std+skew_std+kurt_std,
              subset(data_ids,!is.na(avFD_std)&!is.na(var_std)&
                       !is.na(skew_std)&!is.na(kurt_std))))
```

### Permutational ANOVA

```{r eval=FALSE, include=FALSE}
permANOVA_sel_linear<-permmodels(sel_linear, nperm=10000,type="II",
                                     ncore=16,seed=123)
save(permANOVA_sel_linear,file="output/sel_linear.RData")
```

```{r}
permANOVA_sel_linear$ANOVA
```

### Bootstraped confidence intervals

```{r eval=FALSE, include=FALSE}
b_sel_linear<-bootMer(x=sel_linear,FUN=fixef,nsim=10000)
save(b_sel_linear,file="output/b_sel_linear.RData")
```

```{r}
confint(b_sel_linear,level=0.95, method="boot") # Percentile method
```

### Contrasts for all years

```{r}
emtrends(sel_linear,pairwise~year, var = "avFD_std",adjust="none")$contrasts
emtrends(sel_linear,pairwise~year, var = "var_std",adjust="none")$contrasts
emtrends(sel_linear,pairwise~year, var = "skew_std",adjust="none")$contrasts
emtrends(sel_linear,pairwise~year, var = "kurt_std",adjust="none")$contrasts
```

# Q3: Effects of within-individual variation on fitness components

## Appendix 10: Global models fruit set and proportion of seeds escaping predation

### Fruit set

Using (within-year) standardized traits. Mixed models with id as a random effect.

```{r}
fr_set_linear<-glmer(cbind(fr_set=n_mat_intact_fr,
                           fr_no_set=n_fl-n_mat_intact_fr)~avFD_std+var_std+
                       skew_std+kurt_std+year+ 
                      # Include main effect of year here
                      # but not in selection models!
                      avFD_std:year+var_std:year+skew_std:year+kurt_std:year+
                  (1|id),
                    data_ids,family="binomial",
                  glmerControl(optimizer="bobyqa"))
tab_model(fr_set_linear,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("Fruit set, linear effects"))
Anova(fr_set_linear) 
check_collinearity(glm(cbind(fr_set=n_mat_intact_fr,
                           fr_no_set=n_fl-n_mat_intact_fr)~avFD_std+var_std+
                         skew_std+kurt_std+year,
                       data_ids,family="binomial"))
```

#### ANOVA

```{r}
Anova(fr_set_linear)
```

#### Contrasts for all years

```{r}
emtrends(fr_set_linear,pairwise~year, var = "avFD_std",adjust="none")$contrasts
emtrends(fr_set_linear,pairwise~year, var = "var_std",adjust="none")$contrasts
emtrends(fr_set_linear,pairwise~year, var = "skew_std",adjust="none")$contrasts
emtrends(fr_set_linear,pairwise~year, var = "kurt_std",adjust="none")$contrasts
```

### Proportion of seeds escaping predation

Using (within-year) standardized traits. Mixed models with id as a random effect.

```{r}
data_ids$npr_seeds<-with(data_ids,round(n_seed)-round(n_preyed_seed))
data_ids$pr_seeds<-with(data_ids,round(n_preyed_seed))
seed_escpred_linear<-glmer(cbind(npr_seeds,pr_seeds)~
                             avFD_std+var_std+skew_std+kurt_std+year+ 
                             avFD_std:year+var_std:year+
                             skew_std:year+kurt_std:year+(1|id),
                           subset(data_ids,n_seed>0),family="binomial",
                           glmerControl(optimizer="bobyqa"))
tab_model(seed_escpred_linear,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("Seeds escaping predation, linear effects"))
Anova(seed_escpred_linear)
check_collinearity(glm(cbind(npr_seeds=round(n_seed)-round(n_preyed_seed),
                        pr_seeds=round(n_preyed_seed))~
                          avFD_std+var_std+skew_std+kurt_std+year,
                        data_ids,family="binomial"))
```

According to Anova, differences among years in the effects of avFD and skew.

#### ANOVA

```{r}
Anova(seed_escpred_linear)
```

#### Contrasts for all years

```{r}
emtrends(seed_escpred_linear,pairwise~year, var = "avFD_std",adjust="none")$contrasts
emtrends(seed_escpred_linear,pairwise~year, var = "var_std",adjust="none")$contrasts
emtrends(seed_escpred_linear,pairwise~year, var = "skew_std",adjust="none")$contrasts
emtrends(seed_escpred_linear,pairwise~year, var = "kurt_std",adjust="none")$contrasts
```

## Table 2: Yearly models fitness components

### Fruit set

```{r}
fr_set_linear_87<-glm(cbind(fr_set=n_mat_intact_fr,
                           fr_no_set=n_fl-n_mat_intact_fr)~
                         avFD_std+var_std+skew_std+kurt_std,
                       subset(data_ids,year==1987),family="binomial")
fr_set_linear_88<-glm(cbind(fr_set=n_mat_intact_fr,
                           fr_no_set=n_fl-n_mat_intact_fr)~
                         avFD_std+var_std+skew_std+kurt_std,
                    subset(data_ids,year==1988),family="binomial")
fr_set_linear_89<-glm(cbind(fr_set=n_mat_intact_fr,
                           fr_no_set=n_fl-n_mat_intact_fr)~
                         avFD_std+var_std+skew_std+kurt_std,
                    subset(data_ids,year==1989),family="binomial")
tab_model(fr_set_linear_87,fr_set_linear_88,fr_set_linear_89,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("1987","1988","1989"),
          title="Fruit set, linear effects")
check_collinearity(fr_set_linear_88)
check_collinearity(fr_set_linear_89)
check_collinearity(fr_set_linear_89)
Anova(fr_set_linear_87)
Anova(fr_set_linear_89)
Anova(fr_set_linear_89) 
```

### Proportion of seeds escaping predation

```{r}
seed_escpred_linear_87<-glm(cbind(npr_seeds=round(n_seed)-round(n_preyed_seed),
                                  pr_seeds=round(n_preyed_seed))~
                              avFD_std+var_std+skew_std+kurt_std,
                            subset(data_ids,year==1987&n_seed>0),
                            family="binomial")
data_seed_esc_pred_88<-subset(data_ids,year==1988&n_seed>0&
                                !is.na(avFD_std)&!is.na(var_std)&
                                !is.na(skew_std)&!is.na(kurt_std))
seed_escpred_linear_88<-glm(cbind(npr_seeds=round(n_seed)-round(n_preyed_seed),
                                  pr_seeds=round(n_preyed_seed))~
                              avFD_std+var_std+skew_std+kurt_std,
                            data_seed_esc_pred_88,
                            family="binomial")
seed_escpred_linear_89<-glm(cbind(npr_seeds=round(n_seed)-round(n_preyed_seed),
                                  pr_seeds=round(n_preyed_seed))~
                              avFD_std+var_std+skew_std+kurt_std,
                            subset(data_ids,year==1989&n_seed>0),
                            family="binomial")
tab_model(seed_escpred_linear_87,seed_escpred_linear_88,seed_escpred_linear_89,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("1987","1988","1989"),
          title="Seeds escaping predation, linear effects")
check_collinearity(seed_escpred_linear_87)
check_collinearity(seed_escpred_linear_88)
check_collinearity(seed_escpred_linear_89)
Anova(seed_escpred_linear_87)
Anova(seed_escpred_linear_88)
Anova(seed_escpred_linear_89)
```

##  Appendix 11: Figure effects on fruit set

```{r}
coefs_fr_set_linear<-rbind(
  tidy(fr_set_linear_87)[2:5,]%>%mutate(year=as.factor("1987")),
  tidy(fr_set_linear_88)[2:5,]%>%mutate(year=as.factor("1988")),
  tidy(fr_set_linear_89)[2:5,]%>%mutate(year=as.factor("1989"))
  )%>%
  mutate(signif=as.factor(ifelse(p.value>0.05,0,1)),
         coef=format(round(estimate,3),nsmall=3),
         star=ifelse(signif==1,"*","  "),
         print=paste(coef,star),
         predictor=term)
```

```{r}
prediction_fr_set<-rbind(rbind(tibble(ggpredict(fr_set_linear_87,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(fr_set_linear_88,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(fr_set_linear_89,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="avFD_std"), # avFD_std
                    rbind(tibble(ggpredict(fr_set_linear_87,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(fr_set_linear_88,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(fr_set_linear_89,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="var_std"), # var_std
                    rbind(tibble(ggpredict(fr_set_linear_87,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(fr_set_linear_88,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(fr_set_linear_89,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="skew_std"), # skew_std
                    rbind(tibble(ggpredict(fr_set_linear_87,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(fr_set_linear_88,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(fr_set_linear_89,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="kurt_std"))%>% # kurt_std
  mutate(response="fr_set")%>%
  left_join(coefs_fr_set_linear%>%
              dplyr::select(predictor,coef,year,signif,star,print))%>%
  mutate(predictor=factor(predictor,
                          levels=c("avFD_std","var_std","skew_std","kurt_std")))
```

```{r}
Appendix11<-grid.arrange(
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_fr_set%>%
                  filter(predictor=="avFD_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_fr_set%>%
                filter(predictor=="avFD_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fr_set,avFD_std)%>%
                 pivot_longer(cols=avFD_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("avFD_std"))),
               aes(x=value_predictor,y=fr_set),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dotted", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_fr_set%>%
                  filter(predictor=="var_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_fr_set%>%
                filter(predictor=="var_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fr_set,var_std)%>%
                 pivot_longer(cols=var_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("var_std"))),
               aes(x=value_predictor,y=fr_set),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dotted", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_fr_set%>%
                  filter(predictor=="skew_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_fr_set%>%
                filter(predictor=="skew_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fr_set,skew_std)%>%
                 pivot_longer(cols=skew_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("skew_std"))),
               aes(x=value_predictor,y=fr_set),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dotted", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_fr_set%>%
                  filter(predictor=="kurt_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_fr_set%>%
                filter(predictor=="kurt_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,fr_set,kurt_std)%>%
                 pivot_longer(cols=kurt_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("kurt_std"))),
               aes(x=value_predictor,y=fr_set),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dotted", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ncol=1,
  left=textGrob(
    "Fruit set (proportion of flowers that developed to mature fruits)",
                gp=gpar(fontsize=14,fontfamily="serif"),rot = 90),
  bottom=textGrob("Value of the moment",
                  gp=gpar(fontsize=14,fontfamily="serif")),
  top=textGrob("       1987                                                     1988                                                        1989",
               gp=gpar(fontsize=14,fontfamily="serif"))
  )
ggsave(file="output/figures/Appendix11.tiff",plot=Appendix11,width=10,height=10)
```

## Appendix 12: Figure effects on proportion of seeds escaping predation

```{r}
coefs_seed_escpred_linear<-rbind(
  tidy(seed_escpred_linear_87)[2:5,]%>%mutate(year=as.factor("1987")),
  tidy(seed_escpred_linear_88)[2:5,]%>%mutate(year=as.factor("1988")),
  tidy(seed_escpred_linear_89)[2:5,]%>%mutate(year=as.factor("1989"))
  )%>%
  mutate(signif=as.factor(ifelse(p.value>0.05,0,1)),
         coef=format(round(estimate,3),nsmall=3),
         star=ifelse(signif==1,"*","  "),
         print=paste(coef,star),
         predictor=term)
```

```{r}
prediction_seed_escpred<-rbind(rbind(tibble(ggpredict(seed_escpred_linear_87,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(seed_escpred_linear_88,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(seed_escpred_linear_89,
                                           terms=c("avFD_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="avFD_std"), # avFD_std
                    rbind(tibble(ggpredict(seed_escpred_linear_87,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(seed_escpred_linear_88,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(seed_escpred_linear_89,
                                           terms=c("var_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="var_std"), # var_std
                    rbind(tibble(ggpredict(seed_escpred_linear_87,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(seed_escpred_linear_88,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(seed_escpred_linear_89,
                                           terms=c("skew_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="skew_std"), # skew_std
                    rbind(tibble(ggpredict(seed_escpred_linear_87,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1987)),
                          tibble(ggpredict(seed_escpred_linear_88,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1988)),
                          tibble(ggpredict(seed_escpred_linear_89,
                                           terms=c("kurt_std [all]")))%>%
                            mutate(year=as.factor(1989)))%>%
                      mutate(predictor="kurt_std"))%>% # kurt_std
  mutate(response="seed_escpred")%>%
  left_join(coefs_seed_escpred_linear%>%
              dplyr::select(predictor,coef,year,signif,star,print))%>%
  mutate(predictor=factor(predictor,
                          levels=c("avFD_std","var_std","skew_std","kurt_std")))
```

```{r}
Appendix12<-grid.arrange(
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_seed_escpred%>%
                  filter(predictor=="avFD_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_seed_escpred%>%
                filter(predictor=="avFD_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,prop_seed_preyed,avFD_std)%>%
                 pivot_longer(cols=avFD_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("avFD_std"))),
               aes(x=value_predictor,y=prop_seed_preyed),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dotted", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_seed_escpred%>%
                  filter(predictor=="var_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_seed_escpred%>%
                filter(predictor=="var_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,prop_seed_preyed,var_std)%>%
                 pivot_longer(cols=var_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("var_std"))),
               aes(x=value_predictor,y=prop_seed_preyed),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("solid", "dotted"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_seed_escpred%>%
                  filter(predictor=="skew_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_seed_escpred%>%
                filter(predictor=="skew_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,prop_seed_preyed,skew_std)%>%
                 pivot_longer(cols=skew_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("skew_std"))),
               aes(x=value_predictor,y=prop_seed_preyed),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("solid", "dotted"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ggplot()+
    facet_grid2(c("predictor","year"), 
                labeller=labeller(predictor=predictor.labs), 
                scales = "fixed")+
    geom_ribbon(data=prediction_seed_escpred%>%
                  filter(predictor=="kurt_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                alpha=0.3,fill="#CC0000")+
    geom_line(data=prediction_seed_escpred%>%
                filter(predictor=="kurt_std"),
              aes(x=x,y=predicted,linetype=signif),
              linewidth=1,color="#CC0000")+
    geom_point(data=data_ids%>%
                 dplyr::select(year,prop_seed_preyed,kurt_std)%>%
                 pivot_longer(cols=kurt_std,
                              names_to="predictor",
                              values_to="value_predictor")%>%
                 mutate(predictor=factor(predictor,
                                         levels=c("kurt_std"))),
               aes(x=value_predictor,y=prop_seed_preyed),
               size=2,alpha=0.3)+
    scale_linetype_manual(values=c("dotted", "solid"))+
    ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
  ncol=1,
  left=textGrob(
    "Proportion of seeds escaping predation",
    gp=gpar(fontsize=14,fontfamily="serif"),rot = 90),
  bottom=textGrob("Value of the moment",
                  gp=gpar(fontsize=14,fontfamily="serif")),
  top=textGrob("       1987                                                     1988                                                        1989",
               gp=gpar(fontsize=14,fontfamily="serif"))
)
ggsave(file="output/figures/Appendix12.tiff",plot=Appendix12,width=10,height=10)
```

## Figure 5: Effects on fruit set and proportion of seeds escaping predation

```{r}
data_figure5<-rbind(prediction_fr_set,
                    prediction_seed_escpred)
```

```{r}
legend_fig5<-get_legend(
  ggplot()+
    facet_grid2(c("predictor","year"),
                labeller=labeller(predictor=predictor.labs),
                scales = "free", independent = "x")+
    geom_ribbon(data=data_figure5,
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,fill=response),
                alpha=0.3)+
    geom_line(data=data_figure5,
              aes(x=x,y=predicted,color=response,linetype=signif),linewidth=1)+
    scale_linetype_manual(values=c("dashed", "solid"))+
    scale_color_manual(values=c("#E69F00", "#009E73"),name=NULL,
                       labels=c("Fruit set",
                                "Proportion of seeds escaping predation"))+
    scale_fill_manual(values=c("#E69F00", "#009E73"),name=NULL,
                      labels=c("Fruit set",
                               "Proportion of seeds escaping predation"))+
    ylab("Proportion")+xlab("Value of the moment")+
    my_theme_legend()+theme(legend.position="top")+
    guides(linetype="none")
  )
fig5<-grid.arrange(
  legend_fig5,
  grid.arrange(
    ggplot()+
      facet_grid2(c("predictor","year"), 
                  labeller=labeller(predictor=predictor.labs),scales="fixed")+
      geom_ribbon(data=data_figure5%>%filter(predictor=="avFD_std"),
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                      fill=response),alpha=0.3)+
      geom_line(data=data_figure5%>%filter(predictor=="avFD_std"),
                aes(x=x,y=predicted,color=response,linetype=signif),
                linewidth=1)+
      scale_y_continuous(limits=c(0,1))+
      ylab("Proportion")+xlab("Value of the moment")+
      my_theme_legend()+theme(legend.position="top")+
      scale_fill_manual(values=c("#E69F00", "#009E73"),name=NULL,
                        labels = c("Fruit initiation",
                                   "Proportion of seeds escaping predation"))+
      scale_color_manual(values=c("#E69F00", "#009E73"),name=NULL,
                         labels = c("Fruit initiation",
                                    "Proportion of seeds escaping predation"))+
      scale_linetype_manual(values=c("dashed", "solid"))+
      ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
    ggplot()+
      facet_grid2(c("predictor","year"),
                  labeller=labeller(predictor=predictor.labs),scales="fixed")+
    geom_ribbon(data=data_figure5%>%filter(predictor=="var_std"),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                    fill=response),alpha=0.3)+
      geom_line(data=data_figure5%>%filter(predictor=="var_std"),
                aes(x=x,y=predicted,color=response,linetype=signif),
                linewidth=1)+
      scale_y_continuous(limits=c(0,1))+
      ylab("Proportion")+xlab("Value of the moment")+
      my_theme_legend()+theme(legend.position="top")+
      scale_fill_manual(values=c("#E69F00", "#009E73"),name=NULL,
                        labels = c("Fruit initiation",
                                   "Proportion of seeds escaping predation"))+
    scale_color_manual(values=c("#E69F00", "#009E73"),name=NULL,
                       labels = c("Fruit initiation",
                                  "Proportion of seeds escaping predation"))+
      scale_linetype_manual(values=c("dashed", "solid"))+
      ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
    ggplot()+
      facet_grid2(c("predictor","year"), 
                  labeller=labeller(predictor=predictor.labs),scales="fixed")+
      geom_ribbon(data=data_figure5%>%filter(predictor=="skew_std"),
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                      fill=response),alpha=0.3)+
      geom_line(data=data_figure5%>%filter(predictor=="skew_std"),
                aes(x=x,y=predicted,color=response,linetype=signif),
                linewidth=1)+
      scale_y_continuous(limits=c(0,1))+
      ylab("Proportion")+xlab("Value of the moment")+
      my_theme_legend()+theme(legend.position="top")+
      scale_fill_manual(values=c("#E69F00", "#009E73"),name=NULL,
                        labels = c("Fruit initiation",
                                   "Proportion of seeds escaping predation"))+
      scale_color_manual(values=c("#E69F00", "#009E73"),name=NULL,
                         labels = c("Fruit initiation",
                                    "Proportion of seeds escaping predation"))+
      scale_linetype_manual(values=c("dashed", "solid"))+
      ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
    ggplot()+
      facet_grid2(c("predictor","year"), 
                  labeller=labeller(predictor=predictor.labs),scales="fixed")+
      geom_ribbon(data=data_figure5%>%filter(predictor=="kurt_std"),
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                      fill=response),alpha=0.3)+
      geom_line(data=data_figure5%>%filter(predictor=="kurt_std"),
                aes(x=x,y=predicted,color=response,linetype=signif),
                linewidth=1)+
      scale_y_continuous(limits=c(0,1))+
      ylab("Proportion")+xlab("Value of the moment")+
      my_theme_legend()+theme(legend.position="top")+
      scale_fill_manual(values=c("#E69F00", "#009E73"),name=NULL,
                        labels = c("Fruit initiation",
                                   "Proportion of seeds escaping predation"))+
      scale_color_manual(values=c("#E69F00", "#009E73"),name=NULL,
                         labels = c("Fruit initiation",
                                    "Proportion of seeds escaping predation"))+
      scale_linetype_manual(values=c("dashed", "solid"))+
      ylab(NULL)+xlab(NULL)+my_theme()+theme(strip.text.x = element_blank()),
    ncol=1,
    left=textGrob("Proportion",
                  gp=gpar(fontsize=14,fontfamily="serif"),rot = 90),
    bottom=textGrob("Value of the moment",
                    gp=gpar(fontsize=14,fontfamily="serif")),
    top=textGrob("       1987                                     1988                                           1989",
                 gp=gpar(fontsize=14,fontfamily="serif"))
    ),
  ncol=1,heights=c(0.05,0.95)
  )
ggsave(file="output/figures/fig5.tiff", plot=fig5, width=8, height=10,dpi=300,
       compression="lzw")
```

# Appendix 13: Variation in temperature and number of flowers opening

```{r}
Appendix13<-grid.arrange(
  ggplot(data=data_weather%>%filter(year==1987|year==1988|year==1989)%>%
           filter(month==4|month==5|month==6)%>%
           mutate(yday=yday(date))%>%
           dplyr::select(year,month,date,yday,mean,min,max),
         aes(x=yday,y=mean,ymin=min,ymax=max))+
    facet_grid(cols=vars(year),scales="free")+
    geom_ribbon(fill="lightskyblue1")+geom_line()+my_theme()+
    ylab("Temperature\n(ºC)")+xlab("Day of year")+ggtitle("A)")+
    theme(plot.title = element_text(hjust =-0.09,vjust=-6))+
    geom_vline(xintercept=121,linetype="dashed")+
    geom_vline(xintercept=152,linetype="dashed"),
  ggplot(data=data_weather%>%filter(year==1987|year==1988|year==1989)%>%        
           filter(month==4|month==5|month==6)%>%mutate(yday=yday(date))%>%
           dplyr::select(year,yday,month,mean,min,max)%>%
           group_by(year,month)%>%
           summarise(mean_mean=mean(mean),mean_min=mean(min),mean_max=mean(max),
                     se_mean=std.error(mean),se_min=std.error(min),
                     se_max=std.error(max))%>%
           pivot_longer(-c(year,month),names_to=c("stat","measure"),
                        names_sep="_",values_to="value")%>%
           pivot_wider(names_from="stat",values_from="value")%>%
           filter(measure=="mean"),
         aes(x=year,y=mean,ymin=mean-se,ymax=mean+se,color=factor(month)))+
    geom_jitter(position=position_dodge(width=0.75))+
    geom_errorbar(position=position_dodge(width=0.75),width=0.3)+
    my_theme_legend()+theme(legend.position="top")+
    ylab("Mean temperature (ºC)")+xlab("Year")+ggtitle("B)")+
    theme(plot.title = element_text(hjust =-0.05,vjust=-6))+
    scale_x_continuous(breaks=c(1987,1988,1989))+
    scale_y_continuous(limits=c(2,18))+
    scale_color_discrete(name=NULL,labels=c("April","May","June")),
  ggplot(data=data_id_flowers%>%dplyr::select(id,opening_date_c)%>%
           mutate(year=year(opening_date_c),yday=yday(opening_date_c),
                  date=date(opening_date_c))%>%
           group_by(year,date,yday)%>%
           summarise(n_opening=n()),
         aes(x=yday,y=n_opening))+
    facet_grid(cols=vars(year),scales="free")+
    geom_col(fill="orchid")+my_theme()+
    ylab("Number offlowers\nopening")+xlab(NULL)+ggtitle("C)")+
    theme(plot.title = element_text(hjust =-0.09,vjust=-6))+
    scale_x_continuous(limits=c(90,180))+
    geom_vline(xintercept=121,linetype="dashed")+
    geom_vline(xintercept=152,linetype="dashed"),
  nrow=3
)
ggsave(file="output/figures/Appendix13.tiff", plot=Appendix13,
       width=10, height=12)
ggsave(file="output/figures/Appendix13.svg", plot=Appendix13,
       width=10, height=12) # To add letters for significant differences in B)
```


```{r}
data_weather_8789<-data_weather%>%
  filter(year==1987|year==1988|year==1989)%>%
  filter(month==4|month==5|month==6)%>%
  mutate(year=factor(year))
```

```{r}
model_april_mean<-lm(mean~year,subset(data_weather_8789,month==4))
model_may_mean<-lm(mean~year,subset(data_weather_8789,month==5))
model_june_mean<-lm(mean~year,subset(data_weather_8789,month==6))
```

```{r}
summary(model_april_mean)
Anova(model_april_mean) # No differences among years in April mean temp
summary(model_may_mean)
Anova(model_may_mean) # Differences among years in May mean temp
summary(model_june_mean)
Anova(model_june_mean) # Differences among years in June mean temp
```

```{r}
letters_april_mean<-as.data.frame.list(
  multcompLetters4(aov(mean~year,subset(data_weather_8789,month==4)),
                   TukeyHSD(aov(mean~year,
                                subset(data_weather_8789,month==4))))$year)%>%
  dplyr::select(Letters)%>%
  rownames_to_column(var="year")
letters_may_mean<-as.data.frame.list(
  multcompLetters4(aov(mean~year,subset(data_weather_8789,month==5)),
                   TukeyHSD(aov(mean~year,
                                subset(data_weather_8789,month==5))))$year)%>%
  dplyr::select(Letters)%>%
  rownames_to_column(var="year")
letters_june_mean<-as.data.frame.list(
  multcompLetters4(aov(mean~year,subset(data_weather_8789,month==6)),
                   TukeyHSD(aov(mean~year,
                                subset(data_weather_8789,month==6))))$year)%>%
  dplyr::select(Letters)%>%
  rownames_to_column(var="year")
letters_april_mean
letters_may_mean
letters_june_mean
```

# Session info

```{r}
sessionInfo()
```
